chore: allow the base webmachine resource to be customised
